I18N: Allow for non-ASCII characters in the non-bracketed text form
Modify the s:chars regexps to include non-ASCII characters. Because we only have the endCol of the text form (and cannot easily match beyond that; the line may end there), we cannot simply use strpart(), but have to use matchstr() with /\%c/ to correctly deal with a final non-ASCII character.